[test] Extract CoordinatorEventProcessorTest harness into base class
