+++ fluss-server/src/test/java/org/apache/fluss/server/coordinator/CoordinatorEventProcessorTest.java
@@ -27,7 +27,6 @@ import org.apache.fluss.exception.FencedLeaderEpochException;
 import org.apache.fluss.exception.InvalidAlterTableException;
 import org.apache.fluss.exception.InvalidCoordinatorException;
 import org.apache.fluss.fs.FsPath;
-import org.apache.fluss.metadata.DatabaseDescriptor;
 import org.apache.fluss.metadata.PartitionSpec;
 import org.apache.fluss.metadata.ResolvedPartitionSpec;
 import org.apache.fluss.metadata.Schema;
@@ -62,7 +61,6 @@ import org.apache.fluss.server.coordinator.event.CommitRemoteLogManifestEvent;
 import org.apache.fluss.server.coordinator.event.CoordinatorEventManager;
 import org.apache.fluss.server.coordinator.event.NotifyLeaderAndIsrResponseReceivedEvent;
 import org.apache.fluss.server.coordinator.event.RetryOfflineLeaderEvent;
-import org.apache.fluss.server.coordinator.lease.KvSnapshotLeaseManager;
 import org.apache.fluss.server.coordinator.remote.RemoteDirDynamicLoader;
 import org.apache.fluss.server.coordinator.statemachine.BucketState;
 import org.apache.fluss.server.coordinator.statemachine.ReplicaState;
@@ -81,34 +79,21 @@ import org.apache.fluss.server.metadata.TableMetadata;
 import org.apache.fluss.server.metrics.group.TestingMetricGroups;
 import org.apache.fluss.server.tablet.TestTabletServerGateway;
 import org.apache.fluss.server.zk.NOPErrorHandler;
-import org.apache.fluss.server.zk.ZkEpoch;
 import org.apache.fluss.server.zk.ZooKeeperClient;
-import org.apache.fluss.server.zk.ZooKeeperExtension;
 import org.apache.fluss.server.zk.data.BucketAssignment;
-import org.apache.fluss.server.zk.data.CoordinatorAddress;
 import org.apache.fluss.server.zk.data.LeaderAndIsr;
 import org.apache.fluss.server.zk.data.PartitionAssignment;
 import org.apache.fluss.server.zk.data.TableAssignment;
 import org.apache.fluss.server.zk.data.TabletServerRegistration;
 import org.apache.fluss.server.zk.data.ZkData;
-import org.apache.fluss.server.zk.data.ZkData.PartitionIdsZNode;
-import org.apache.fluss.server.zk.data.ZkData.TableIdsZNode;
 import org.apache.fluss.server.zk.data.ZkVersion;
-import org.apache.fluss.testutils.common.AllCallbackWrapper;
 import org.apache.fluss.testutils.common.ManuallyTriggeredScheduledExecutorService;
 import org.apache.fluss.types.DataTypes;
 import org.apache.fluss.utils.ExceptionUtils;
 import org.apache.fluss.utils.clock.SystemClock;
-import org.apache.fluss.utils.concurrent.ExecutorThreadFactory;
-import org.apache.fluss.utils.concurrent.FlussScheduler;
-import org.apache.fluss.utils.concurrent.Scheduler;
 import org.apache.fluss.utils.types.Tuple2;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.io.TempDir;
 
 import java.nio.file.Path;
@@ -124,7 +109,6 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
@@ -152,7 +136,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /** Test for {@link CoordinatorEventProcessor}. */
-class CoordinatorEventProcessorTest {
+class CoordinatorEventProcessorTest extends CoordinatorEventProcessorTestBase {
 
     private static final int N_BUCKETS = 3;
     private static final int REPLICATION_FACTOR = 3;
@@ -168,107 +152,6 @@ class CoordinatorEventProcessorTest {
                     .property(ConfigOptions.TABLE_KV_STANDBY_REPLICA_ENABLED.key(), "true")
                     .build()
                     .withReplicationFactor(REPLICATION_FACTOR);
-
-    @RegisterExtension
-    public static final AllCallbackWrapper<ZooKeeperExtension> ZOO_KEEPER_EXTENSION_WRAPPER =
-            new AllCallbackWrapper<>(new ZooKeeperExtension());
-
-    private static ZooKeeperClient zookeeperClient;
-    private static MetadataManager metadataManager;
-    private static ZkEpoch zkEpoch;
-
-    private CoordinatorEventProcessor eventProcessor;
-    private final String defaultDatabase = "db";
-    private TestCoordinatorChannelManager testCoordinatorChannelManager;
-    private AutoPartitionManager autoPartitionManager;
-    private LakeTableTieringManager lakeTableTieringManager;
-    private CompletedSnapshotStoreManager completedSnapshotStoreManager;
-    private CoordinatorMetadataCache serverMetadataCache;
-    private ReplicaCapacityController replicaCapacityController;
-    private KvSnapshotLeaseManager kvSnapshotLeaseManager;
-    private Scheduler scheduler;
-    private String remoteDataDir;
-
-    @BeforeAll
-    static void baseBeforeAll() throws Exception {
-        zookeeperClient =
-                ZOO_KEEPER_EXTENSION_WRAPPER
-                        .getCustomExtension()
-                        .getZooKeeperClient(NOPErrorHandler.INSTANCE);
-        metadataManager =
-                new MetadataManager(
-                        zookeeperClient,
-                        new Configuration(),
-                        new LakeCatalogDynamicLoader(new Configuration(), null, true));
-
-        // register coordinator server
-        zookeeperClient.registerCoordinatorLeader(
-                new CoordinatorAddress(
-                        "2", Endpoint.fromListenersString("CLIENT://localhost:10012")));
-
-        zkEpoch = zookeeperClient.fenceBecomeCoordinatorLeader("2");
-        // register 3 tablet servers
-        for (int i = 0; i < 3; i++) {
-            zookeeperClient.registerTabletServer(
-                    i,
-                    new TabletServerRegistration(
-                            "rack" + i,
-                            Collections.singletonList(
-                                    new Endpoint("host" + i, 1000, DEFAULT_LISTENER_NAME)),
-                            System.currentTimeMillis()));
-        }
-    }
-
-    @BeforeEach
-    void beforeEach() {
-        serverMetadataCache = new CoordinatorMetadataCache();
-        // set a test channel manager for the context
-        testCoordinatorChannelManager = new TestCoordinatorChannelManager();
-        lakeTableTieringManager =
-                new LakeTableTieringManager(TestingMetricGroups.LAKE_TIERING_METRICS);
-        remoteDataDir = zookeeperClient.getDefaultRemoteDataDir();
-        Configuration conf = new Configuration();
-        conf.setString(ConfigOptions.REMOTE_DATA_DIR, remoteDataDir);
-        replicaCapacityController = new ReplicaCapacityController(conf, serverMetadataCache);
-        autoPartitionManager =
-                new AutoPartitionManager(
-                        serverMetadataCache,
-                        metadataManager,
-                        new RemoteDirDynamicLoader(conf),
-                        conf,
-                        replicaCapacityController);
-        kvSnapshotLeaseManager =
-                new KvSnapshotLeaseManager(
-                        Duration.ofMinutes(10).toMillis(),
-                        zookeeperClient,
-                        remoteDataDir,
-                        SystemClock.getInstance(),
-                        TestingMetricGroups.COORDINATOR_METRICS);
-        kvSnapshotLeaseManager.start();
-
-        scheduler = new FlussScheduler(1);
-        scheduler.startup();
-
-        eventProcessor = buildCoordinatorEventProcessor();
-        eventProcessor.startup();
-        metadataManager.createDatabase(
-                defaultDatabase, DatabaseDescriptor.builder().build(), false);
-        completedSnapshotStoreManager = eventProcessor.completedSnapshotStoreManager();
-    }
-
-    @AfterEach
-    void afterEach() throws Exception {
-        if (eventProcessor != null) {
-            eventProcessor.shutdown();
-        }
-        if (scheduler != null) {
-            scheduler.shutdown();
-        }
-        metadataManager.dropDatabase(defaultDatabase, false, true);
-        // clear the assignment info for all tables;
-        ZOO_KEEPER_EXTENSION_WRAPPER.getCustomExtension().cleanupPath(TableIdsZNode.path());
-        ZOO_KEEPER_EXTENSION_WRAPPER.getCustomExtension().cleanupPath(PartitionIdsZNode.path());
-    }
 
     @Test
     void testLoadedAssignmentsTrackKnownKvAndUnknownTablesConservatively() throws Exception {
@@ -2430,27 +2313,6 @@ class CoordinatorEventProcessorTest {
         assertThat(leaderAndIsr.isr())
                 .isEqualTo(newLeaderAndIsrOfZk.isr())
                 .hasSameElementsAs(expectedIsr);
-    }
-
-    private CoordinatorEventProcessor buildCoordinatorEventProcessor() {
-        Configuration conf = new Configuration();
-        conf.set(ConfigOptions.REMOTE_DATA_DIR, remoteDataDir);
-        conf.set(ConfigOptions.COORDINATOR_OFFLINE_LEADER_RETRY_DELAY, Duration.ofDays(1));
-        return new CoordinatorEventProcessor(
-                zookeeperClient,
-                serverMetadataCache,
-                testCoordinatorChannelManager,
-                new CoordinatorContext(zkEpoch),
-                replicaCapacityController,
-                autoPartitionManager,
-                lakeTableTieringManager,
-                TestingMetricGroups.COORDINATOR_METRICS,
-                conf,
-                Executors.newFixedThreadPool(1, new ExecutorThreadFactory("test-coordinator-io")),
-                metadataManager,
-                kvSnapshotLeaseManager,
-                scheduler,
-                SystemClock.getInstance());
     }
 
     private static class FailingUpdateMetadataChannelManager extends TestCoordinatorChannelManager {
